+++ prime_ai_trader/priceaction/__init__.py
@@ -1,4 +1,11 @@
+from .candles import (
+    CandlestickAssessment, CandlestickPattern, analyze_candlestick_patterns,
+    candlestick_feature_frame,
+)
 from .structure import analyze_structure, detect_pivots, support_resistance_zones
 
-__all__ = ["analyze_structure", "detect_pivots", "support_resistance_zones"]
-
+__all__ = [
+    "CandlestickAssessment", "CandlestickPattern", "analyze_candlestick_patterns",
+    "candlestick_feature_frame", "analyze_structure", "detect_pivots",
+    "support_resistance_zones",
+]